Move post-update notification logic to UpdateManager base class
